+++ test/test.cpp
@@ -228,7 +228,8 @@ void test_Machine() {
 	ASSERTEQUALS(JT("{'dh':1,'en':1,'ho':0,'is':0,'mi':16,'sa':1.8,'tm':32000,'tn':-32000,'ud':0}"), buf);
 	ASSERTEQUAL((size_t)(void*)out, (size_t)(void*)buf+strlen(buf));
 	out = machine.saveSysConfig(buf, sizeof(buf));
-	ASSERTEQUALS(JT("{'ah':0,'as':0,'db':0,'eu':2000,'hp':3,'jp':0,'lb':200,'lh':0,"
+#define HASH1 "2144573825"
+	ASSERTEQUALS(JT("{'ah':0,'as':0,'ch':" HASH1 ",'db':0,'hp':3,'jp':0,'lb':200,'lh':0,"
 				 "'mv':12800,'om':0,'pc':2,'pi':11,'to':0,'tv':0.70}"), 
 				 buf);
 	ASSERTEQUAL((size_t)(void*)out, (size_t)(void*)buf+strlen(buf));
@@ -730,7 +731,7 @@ void test_JsonController() {
     jc.process(jcmd);
     char sysbuf[500];
     const char *fmt = "{'s':%d,'r':{'sys':"\
-                      "{'ah':false,'as':false,'ch':2144574995,'eu':2000,'fr':1000,'hp':3,'jp':false,'lb':200,'lh':false,"\
+                      "{'ah':false,'as':false,'ch':2144573891,'fr':1000,'hp':3,'jp':false,'lb':200,'lh':false,"\
                       "'lp':0,'mv':12800,'om':0,'pc':2,'pi':11,'sd':800,'tc':12345,"\
                       "'to':0,'tv':0.700,'v':%.3f}"\
                       "},'t':0.000}\n";
@@ -1949,7 +1950,7 @@ void test_autoSync() {
 
     ASSERTEQUAL(1000, MAX_JSON);
     uint8_t *eeaddr = 0;
-    ASSERTEQUAL(0, mt.syncHash);
+    ASSERTEQUAL(0, machine.syncHash);
     ASSERT(!machine.autoSync);
     ASSERTEQUAL(STATUS_BUSY_SETUP, mt.status);
     mt.loop();
@@ -1957,8 +1958,8 @@ void test_autoSync() {
     mt.loop();
     ASSERTEQUAL(STATUS_WAIT_IDLE, mt.status);
     Serial.clear(); // banner
-    ASSERT(mt.syncHash);
-    uint32_t hash1 = mt.syncHash;
+    ASSERT(machine.syncHash);
+    uint32_t hash1 = machine.syncHash;
 	ASSERTEQUAL(hash1, machine.hash());
 	machine.axis[4].enable(false);
 	machine.axis[5].enable(false); 
@@ -1975,14 +1976,17 @@ void test_autoSync() {
 		Serial.output().c_str());
     ASSERTEQUAL(STATUS_OK, mt.status);
     ASSERT(machine.autoSync);
-	ASSERT(hash2 != machine.hash());
+	int32_t hash3 = machine.hash();
+	ASSERT(hash2 != hash3);
 	ASSERTEQUAL(false, machine.axis[4].isEnabled());
+#define HASH3 "2146409185"
+	snprintf(buf, sizeof(buf), "%ld", (long) hash3);
+	ASSERTEQUALS(HASH3, buf);
 
     mt.loop();
-#define HASH_EXPECTED "2146409777"
+	string eeprom3 = eeprom_read_string(0);
     ASSERTEQUALS(JT( "["
-					 "{'cmt':'sysch:" HASH_EXPECTED "'},"
-                     "{'sys':{'ah':1,'as':1,'db':0,'eu':2000,'hp':3,'jp':0,"
+                     "{'sys':{'ah':1,'as':1,'ch':" HASH3 ",'db':0,'hp':3,'jp':0,"
 					 "'lb':200,'lh':0,'mv':12800,'om':3,'pc':2,'pi':11,'to':0,'tv':0.70}},"
                      "{'x':{'dh':1,'en':1,'ho':0,'is':0,'mi':16,'sa':1.8,'tm':32000,'tn':-32000,'ud':0}},"
                      "{'y':{'dh':1,'en':1,'ho':0,'is':0,'mi':16,'sa':1.8,'tm':32000,'tn':-32000,'ud':0}},"
@@ -1991,19 +1995,21 @@ void test_autoSync() {
                      "{'b':{'en':0}},"
                      "{'c':{'en':0}},"
 					 "{'hom':''}]"),
-                 eeprom_read_string(0).c_str());
+                 eeprom3.c_str());
+	snprintf(buf, sizeof(buf), "%ld", (long) machine.syncHash);
+	ASSERTEQUALS(HASH3, buf);
     ASSERTEQUAL(STATUS_WAIT_IDLE, mt.status);
     mt.loop();
-    ASSERT(hash1 != mt.syncHash);
+    ASSERT(hash1 != machine.syncHash);
     mt.loop();
     ASSERTEQUAL(STATUS_WAIT_IDLE, mt.status);
-    uint32_t hash3 = mt.syncHash;
     ASSERTEQUAL(hash3, machine.hash());
+    ASSERTEQUAL(hash3, machine.syncHash);
     mt.loop();
     ASSERTEQUAL(STATUS_WAIT_IDLE, mt.status);
 
 	// simulate restart
-	mt.syncHash = 0;
+	machine.syncHash = 0;
 	mt.printBannerOnIdle = true;
 	machine.autoSync = false;
 	mt.status = STATUS_BUSY_SETUP;
@@ -2012,15 +2018,12 @@ void test_autoSync() {
     mt.loop(); // parse startup JSON
 	Serial.clear(); // banner
     ASSERTEQUAL(STATUS_BUSY_PARSED, mt.status);
-    mt.loop(); // CMT
-    ASSERTEQUAL(STATUS_BUSY_PARSED, mt.status);
-    ASSERTEQUALS(JT("sysch:" HASH_EXPECTED "\n"
-					"{'s':0,'r':{'cmt':'sysch:" HASH_EXPECTED "'},'t':0.000}\n"), Serial.output().c_str());
     mt.loop(); // sys
     ASSERTEQUAL(STATUS_BUSY_PARSED, mt.status);
 	ASSERTEQUAL(true, machine.autoSync);
+	ASSERTEQUALS(eeprom3.c_str(), eeprom_read_string(0).c_str());
 	ASSERTEQUALS(JT("{'s':0,'r':" 
-                    "{'sys':{'ah':true,'as':true,'db':0,'eu':2000,'hp':3,"
+                    "{'sys':{'ah':true,'as':true,'ch':" HASH3 ",'db':0,'hp':3,"
 					"'jp':false,'lb':200,'lh':false,'mv':12800,'om':3,'pc':2,'pi':11,'to':0,'tv':0.700}},"
 					"'t':0.000}\n"),
 					Serial.output().c_str());
@@ -2055,7 +2058,7 @@ void test_autoSync() {
     mt.loop(); // banner
     ASSERTEQUAL(STATUS_WAIT_IDLE, mt.status);
 	snprintf(buf, sizeof(buf), "FireStep %d.%d.%d sysch:%s\n", 
-		VERSION_MAJOR, VERSION_MINOR, VERSION_PATCH, HASH_EXPECTED);
+		VERSION_MAJOR, VERSION_MINOR, VERSION_PATCH, HASH3);
 	ASSERTEQUALS(buf, Serial.output().c_str());
 
     cout << "TEST	: test_autoSync() OK " << endl;